compute all FA-infer tiling params and mask on AICPU
