+++ csrc/flash_attn_npu_v3/tilingdata.h
@@ -58,6 +58,7 @@ struct FAInferTilingData {
     uint64_t splitOTotalSize;
     uint32_t totalSplitNodeNum;
     uint32_t needCoreNum;
+    uint32_t flashDecodeFlag;
     uint32_t numSplits;
     coreNode coreInfo[25];
     splitNode splitInfo[25];
@@ -88,6 +89,7 @@ struct FAInferTilingData {
     uint64_t get_splitOTotalSize() const { return splitOTotalSize; }
     uint32_t get_totalSplitNodeNum() const { return totalSplitNodeNum; }
     uint32_t get_needCoreNum() const { return needCoreNum; }
+    uint32_t get_flashDecodeFlag() const { return flashDecodeFlag; }
     uint32_t get_numSplits() const { return numSplits; }
 
     void set_numHeads(uint32_t value) { numHeads = value; }
@@ -116,6 +118,7 @@ struct FAInferTilingData {
     void set_splitOTotalSize(uint64_t value) { splitOTotalSize = value; }
     void set_totalSplitNodeNum(uint32_t value) { totalSplitNodeNum = value; }
     void set_needCoreNum(uint32_t value) { needCoreNum = value; }
+    void set_flashDecodeFlag(uint32_t value) { flashDecodeFlag = value; }
     void set_numSplits(uint32_t value) { numSplits = value; }
 };
