+++ flash_attn_npu_v3/__init__.py
@@ -9,4 +9,5 @@ else:
         flash_attn_func,
         flash_attn_varlen_func,
         flash_attn_with_kvcache,
+        get_scheduler_metadata,
     )
